move catchup files to separate directory
