+++ src/catchup/VerifyLedgerChainWork.cpp
@@ -2,8 +2,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "historywork/VerifyLedgerChainWork.h"
-#include "history/CatchupManager.h"
+#include "catchup/VerifyLedgerChainWork.h"
 #include "history/FileTransferInfo.h"
 #include "historywork/Progress.h"
 #include "ledger/LedgerHeaderFrame.h"
