+++ src/historywork/BatchDownloadWork.cpp
@@ -3,7 +3,7 @@
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "historywork/BatchDownloadWork.h"
-#include "history/CatchupManager.h"
+#include "catchup/CatchupManager.h"
 #include "history/HistoryManager.h"
 #include "historywork/GetAndUnzipRemoteFileWork.h"
 #include "historywork/Progress.h"
