+++ src/main/main.cpp
@@ -4,11 +4,11 @@
 #include "util/asio.h"
 #include "StellarCoreVersion.h"
 #include "bucket/Bucket.h"
+#include "catchup/CatchupManager.h"
 #include "crypto/Hex.h"
 #include "crypto/KeyUtils.h"
 #include "crypto/SecretKey.h"
 #include "database/Database.h"
-#include "history/CatchupManager.h"
 #include "history/HistoryManager.h"
 #include "ledger/LedgerManager.h"
 #include "lib/http/HttpClient.h"
